Use photo array from front matter; new summary
As I move to Micropub for creating content, I can simply ensure that
uploaded images are references in the front matter of each post.  Then
the Hugo template can insert an `<img>` tag, complete with alt text,
into the body of the post.  I can still add additional body text, if I
so desire.

Update the generation of summary content. If a `description` is present
in the front matter, use that.  If not, and there is content, use the
[automatic
summary](https://gohugo.io/content-management/summaries/#hugo-defined-automatic-summary-splitting)
generated by Hugo.  If there is no content (and thus no summary), then
use the title.