Add photo support, with shortcode
If the front matter contains a "photos" array, allow these photos to be
inserted by means of the "photo" shortcode.
```
{{< photo 1 "alternate text">}}
```
This would add the second photo in the array into the post, with the
desginated alternate text.  (Remember: arrays are zero-indexed!)

If there is a photos array in the front matter, and the `photos`
shortcode is NOT invoked, then add the first (zeroth) photo from the
array to the front of the content, as a featured image of sorts.