Back to partials!
I didn't like storing marked-up repost and reply content in the body of
my content. So I've moved it to the front matter.  In this way, I can
mark it up when I render the site, and therefore more easily change the
markup over time without editing every single post.